build: improve detection of endpoints which support pagination
This improves the logic we use to figure out which API endpoints
support pagination. Today, we look for an ID which begins with
`list`, but it turns out that many endpoints with pagination
support don't follow this convention.

This instead switches to a slightly more complex hybrid approach.
We rely on looking for the `per_page` query parameter, which is
a good sign of pagination support. On top of that, we hardcode:

* (a) some operations which support the `per_page` parameter in
  the code, but where the OpenAPI specification is wrong
* (b) some operations with support the `per_page` paraneter but
  do pagination differently
* (c) some (actually one!) operations which still paginate in
  the way that this plugin supports (`Link` header discovery),
  but use different params

The cases for (a) have been flagged to the teams at GitHub.
To better solve (b) and (c), we will look at adding some
explicit pagination information to the schema.